move tag/snap logic from consumer to snapping component, move merge-command to a new aspect "merging"
