+++ .bitmap
@@ -561,6 +561,13 @@
         "mainFile": "index.ts",
         "rootDir": "scopes/mdx/mdx"
     },
+    "merging": {
+        "scope": "",
+        "version": "",
+        "defaultScope": "teambit.component",
+        "mainFile": "index.ts",
+        "rootDir": "scopes/component/merging"
+    },
     "mocha": {
         "scope": "teambit.defender",
         "version": "0.0.19",
