+++ scopes/component/merging/index.ts
@@ -1,0 +1,5 @@
+import { MergingAspect } from './merging.aspect';
+
+export type { MergingMain } from './merging.main.runtime';
+export default MergingAspect;
+export { MergingAspect };
