+++ scopes/component/merging/merging.aspect.ts
@@ -1,0 +1,5 @@
+import { Aspect } from '@teambit/harmony';
+
+export const MergingAspect = Aspect.create({
+  id: 'teambit.component/merging',
+});
