+++ scopes/harmony/bit/manifests.ts
@@ -80,6 +80,7 @@ import { BitCustomAspectAspect } from '@teambit/bit-custom-aspect';
 import { CommunityAspect } from '@teambit/community';
 import { CloudAspect } from '@teambit/cloud';
 import { SnappingAspect } from '@teambit/snapping';
+import { MergingAspect } from '@teambit/merging';
 import { BitAspect } from './bit.aspect';
 
 export const manifestsMap = {
@@ -164,6 +165,7 @@ export const manifestsMap = {
   [CommunityAspect.id]: CommunityAspect,
   [CloudAspect.id]: CloudAspect,
   [SnappingAspect.id]: SnappingAspect,
+  [MergingAspect.id]: MergingAspect,
 };
 
 export function isCoreAspect(id: string) {
